Concrete resolution: add missing linear constraints
This fixes the following bug: imagine an unknown pword c
with |c.u| = 2 and max_burst = 1. It was possible to
get I_c(2) = 1 from the linear solver because the I_c(2)
linear unknown was not constrained w.r.t max_burst.

We now generate the constraint:
	I_c(j) >= j / max_burst
for the smallest j appearing in the linear system.